fix(forms): refactored forms component. It is now an organism instead of a molecule
