+++ src/components/organisms/forms/forms.stories.mdx
@@ -2,7 +2,7 @@ import {Meta, Preview, Story} from "@storybook/addon-docs/blocks";
 import {SignInForm} from "./signInForm";
 import {action} from "@storybook/addon-actions";
 
-<Meta title="Molecules/Forms" component={SignInForm}/>
+<Meta title="Organisms/Forms" component={SignInForm}/>
 
 # Forms
