+++ src/components/organisms/forms/signInForm.js
@@ -2,14 +2,14 @@ import React from "react";
 import {PrimaryButton, SecondaryButton} from "../../atoms/button/buttons";
 import styled from "styled-components";
 import {typeScale} from "../../../utils";
-import {EmailInput, PasswordInput} from "../textFields/textFields";
+import {EmailInput, PasswordInput} from "../../molecules/textFields/textFields";
 
 const Title = styled.h3`
   font-size: ${typeScale.header3};
 `;
 
 export const SignInForm = ({title = 'form', signIn, signUp}) => (
-  <div>
+  <form>
     <Title>{title}</Title>
     <EmailInput label="Email" placeholder="yeny.garcia@gmail.com" />
     <PasswordInput label="Password" />
@@ -19,5 +19,5 @@ export const SignInForm = ({title = 'form', signIn, signUp}) => (
     <PrimaryButton onClick={signIn}>
         Sign In
     </PrimaryButton>
-  </div>
+  </form>
 );
